+++ command/agent/config_test.go
@@ -104,15 +104,15 @@ func TestDecodeConfig(t *testing.T) {
 		t.Fatalf("bad: %#v", config)
 	}
 
-	if config.Protocol != DefaultConfig.Protocol {
+	if config.Protocol != DefaultConfig().Protocol {
 		t.Fatalf("bad: %#v", config)
 	}
 
-	if config.SkipLeaveOnInt != DefaultConfig.SkipLeaveOnInt {
+	if config.SkipLeaveOnInt != DefaultConfig().SkipLeaveOnInt {
 		t.Fatalf("bad: %#v", config)
 	}
 
-	if config.LeaveOnTerm != DefaultConfig.LeaveOnTerm {
+	if config.LeaveOnTerm != DefaultConfig().LeaveOnTerm {
 		t.Fatalf("bad: %#v", config)
 	}
 
@@ -172,6 +172,20 @@ func TestDecodeConfig(t *testing.T) {
 	}
 
 	if config.SkipLeaveOnInt != true {
+		t.Fatalf("bad: %#v", config)
+	}
+
+	// tags
+	input = `{"tags": {"foo": "bar", "role": "test"}}`
+	config, err = DecodeConfig(bytes.NewReader([]byte(input)))
+	if err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	if config.Tags["foo"] != "bar" {
+		t.Fatalf("bad: %#v", config)
+	}
+	if config.Tags["role"] != "test" {
 		t.Fatalf("bad: %#v", config)
 	}
 }